pipeline_components/vlm_testing/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
tqdm>=4.65.0
pandas>=1.5.0
matplotlib>=3.7.0
scikit-learn>=1.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
opencv-python>=4.7.0
diffusers>=0.14.0
safetensors>=0.3.1
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
tabulate>=0.9.0